move geo scalar functions to io/
