+++ src/algorithm/geo/contains.rs
@@ -1,12 +1,12 @@
-use crate::algorithm::geo::utils::{
-    geometry_collection_to_geo, geometry_to_geo, line_string_to_geo, multi_line_string_to_geo,
-    multi_point_to_geo, multi_polygon_to_geo, point_to_geo, polygon_to_geo,
-};
 use crate::array::*;
 use crate::datatypes::GeoDataType;
 use crate::geo_traits::{
     GeometryCollectionTrait, GeometryTrait, LineStringTrait, MultiLineStringTrait, MultiPointTrait,
     MultiPolygonTrait, PointTrait, PolygonTrait,
+};
+use crate::io::geo::scalar::{
+    geometry_collection_to_geo, geometry_to_geo, line_string_to_geo, multi_line_string_to_geo,
+    multi_point_to_geo, multi_polygon_to_geo, point_to_geo, polygon_to_geo,
 };
 use crate::GeometryArrayTrait;
 use arrow_array::builder::BooleanBuilder;
