+++ src/scalar/binary/scalar.rs
@@ -1,4 +1,4 @@
-use crate::algorithm::geo::utils::geometry_to_geo;
+use crate::io::geo::scalar::geometry_to_geo;
 use crate::trait_::GeometryScalarTrait;
 use arrow_array::{GenericBinaryArray, OffsetSizeTrait};
 use geo::BoundingRect;
