+++ src/scalar/coord/interleaved/scalar.rs
@@ -1,9 +1,9 @@
 use arrow_buffer::ScalarBuffer;
 use rstar::{RTreeObject, AABB};
 
-use crate::algorithm::geo::utils::coord_to_geo;
 use crate::algorithm::native::eq::coord_eq;
 use crate::geo_traits::CoordTrait;
+use crate::io::geo::scalar::coord_to_geo;
 use crate::scalar::SeparatedCoord;
 use crate::trait_::GeometryScalarTrait;
