+++ src/scalar/coord/separated/scalar.rs
@@ -1,6 +1,6 @@
-use crate::algorithm::geo::utils::coord_to_geo;
 use crate::algorithm::native::eq::coord_eq;
 use crate::geo_traits::CoordTrait;
+use crate::io::geo::scalar::coord_to_geo;
 use crate::scalar::InterleavedCoord;
 use crate::trait_::GeometryScalarTrait;
 use arrow_buffer::ScalarBuffer;
