+++ src/scalar/linestring/scalar.rs
@@ -1,9 +1,9 @@
-use crate::algorithm::geo::utils::line_string_to_geo;
 use crate::algorithm::native::bounding_rect::bounding_rect_linestring;
 use crate::algorithm::native::eq::line_string_eq;
 use crate::array::util::OffsetBufferUtils;
 use crate::array::{CoordBuffer, LineStringArray};
 use crate::geo_traits::LineStringTrait;
+use crate::io::geo::scalar::line_string_to_geo;
 use crate::scalar::Point;
 use crate::trait_::{GeometryArraySelfMethods, GeometryScalarTrait};
 use arrow_array::OffsetSizeTrait;
