+++ src/scalar/multipoint/scalar.rs
@@ -1,9 +1,9 @@
-use crate::algorithm::geo::utils::multi_point_to_geo;
 use crate::algorithm::native::bounding_rect::bounding_rect_multipoint;
 use crate::algorithm::native::eq::multi_point_eq;
 use crate::array::util::OffsetBufferUtils;
 use crate::array::{CoordBuffer, MultiPointArray};
 use crate::geo_traits::MultiPointTrait;
+use crate::io::geo::scalar::multi_point_to_geo;
 use crate::scalar::Point;
 use crate::trait_::GeometryArraySelfMethods;
 use crate::trait_::GeometryScalarTrait;
