+++ src/scalar/point/scalar.rs
@@ -1,8 +1,8 @@
-use crate::algorithm::geo::utils::{coord_to_geo, point_to_geo};
 use crate::algorithm::native::bounding_rect::bounding_rect_point;
 use crate::algorithm::native::eq::point_eq;
 use crate::array::CoordBuffer;
 use crate::geo_traits::{CoordTrait, PointTrait};
+use crate::io::geo::scalar::{coord_to_geo, point_to_geo};
 use crate::trait_::{GeometryArraySelfMethods, GeometryScalarTrait};
 use rstar::{RTreeObject, AABB};
 use std::borrow::Cow;
