+++ src/scalar/polygon/scalar.rs
@@ -1,9 +1,9 @@
-use crate::algorithm::geo::utils::polygon_to_geo;
 use crate::algorithm::native::bounding_rect::bounding_rect_polygon;
 use crate::algorithm::native::eq::polygon_eq;
 use crate::array::util::OffsetBufferUtils;
 use crate::array::{CoordBuffer, PolygonArray};
 use crate::geo_traits::PolygonTrait;
+use crate::io::geo::scalar::polygon_to_geo;
 use crate::scalar::LineString;
 use crate::trait_::{GeometryArraySelfMethods, GeometryScalarTrait};
 use arrow_array::OffsetSizeTrait;
